Removing most methods from CatalogManager
